usecases/ai/rag-toolkit/backend/requirements.txt
# Serving
fastapi[all]==0.111.0
sse_starlette==2.1.2

# Utils
scipy
soundfile
numpy==1.26.4
openai==1.56.2
pyyaml==6.0.1
pypdf==5.0.0
langchain==0.3.8
langchain-chroma==0.1.4
langchain-community===0.3.8
chromadb==0.5.18
huggingface_hub>=0.23.0
botocore==1.34.88
cached_path==1.6.3

# OpenVINO
--extra-index-url https://download.pytorch.org/whl/cpu
torch==2.4.0
torchaudio==2.4.0
openvino==2024.6.0
optimum-intel[openvino,nncf]==1.21.0
